srs_analysis.py
from docx import Document

def analyze_srs(file_path):
    # For testing, we can return dummy data instead of reading the actual document
    extracted_data = {
        "ui_components": ["Button", "Form", "Table"],
        "state_management": "Global state management required",
        "api_endpoints": [
            "GET /api/dashboard",
            "POST /api/lms/leave/apply"
        ],
        "accessibility_requirements": "All components must be accessible",
        "styling_guidelines": {
            "primary_color": "#007bff",
            "font_family": "Inter"
        }
    }
    
    # Uncomment the following lines if you want to read from the document
    # doc = Document(file_path)
    # ... (existing extraction logic)

    return extracted_data
