requirements.txt
requests
python-docx
python-dotenv
langchain
langchain-groq
